intermediate commit
continue creating basement of the system

+++ src/consts.h
@@ -3,11 +3,13 @@
 
 #include <limits>
 #include <string>
+#include <vector>
+#include <assert.h>
 
 namespace consts {
   static constexpr double undefined_double = std::numeric_limits<double>::max();
   /// ms smallest unit of time, each base component func must durate that value.
-  static constexpr double atomic_time_value = 500;
+  static constexpr unsigned int atomic_time_value = 500;
   static constexpr char function_manager_name[] = "function manager";
 };
